fix: resolve compiler warnings in format strings and buffer sizes
Fix format specifier mismatches for unsigned types and increase
buffer sizes to prevent potential overflows. Use PRIu64 macro for
uint64_t timestamps and add explicit casts for enum logging.

+++ src/components/ble/SimpleWeatherService.cpp
@@ -20,6 +20,7 @@
 
 #include <algorithm>
 #include <array>
+#include <cinttypes>
 #include <cstring>
 #include <nrf_log.h>
 
@@ -132,12 +133,12 @@ int SimpleWeatherService::OnCommand(struct ble_gatt_access_ctxt* ctxt) {
     case MessageType::CurrentWeather:
       if (GetVersion(dataBuffer) <= 1) {
         currentWeather = CreateCurrentWeather(dataBuffer);
-        NRF_LOG_INFO("Current weather :\n\tTimestamp : %d\n\tTemperature:%d\n\tMin:%d\n\tMax:%d\n\tIcon:%d\n\tLocation:%s",
+        NRF_LOG_INFO("Current weather :\n\tTimestamp : %" PRIu64 "\n\tTemperature:%d\n\tMin:%d\n\tMax:%d\n\tIcon:%d\n\tLocation:%s",
                      currentWeather->timestamp,
                      currentWeather->temperature.PreciseCelsius(),
                      currentWeather->minTemperature.PreciseCelsius(),
                      currentWeather->maxTemperature.PreciseCelsius(),
-                     currentWeather->iconId,
+                     static_cast<int>(currentWeather->iconId),
                      currentWeather->location.data());
         if (GetVersion(dataBuffer) == 1) {
           NRF_LOG_INFO("Sunrise: %d\n\tSunset: %d", currentWeather->sunrise, currentWeather->sunset);
@@ -147,13 +148,13 @@ int SimpleWeatherService::OnCommand(struct ble_gatt_access_ctxt* ctxt) {
     case MessageType::Forecast:
       if (GetVersion(dataBuffer) == 0) {
         forecast = CreateForecast(dataBuffer);
-        NRF_LOG_INFO("Forecast : Timestamp : %d", forecast->timestamp);
+        NRF_LOG_INFO("Forecast : Timestamp : %" PRIu64, forecast->timestamp);
         for (int i = 0; i < 5; i++) {
           NRF_LOG_INFO("\t[%d] Min: %d - Max : %d - Icon : %d",
                        i,
                        forecast->days[i]->minTemperature.PreciseCelsius(),
                        forecast->days[i]->maxTemperature.PreciseCelsius(),
-                       forecast->days[i]->iconId);
+                       static_cast<int>(forecast->days[i]->iconId));
         }
       }
       break;
